Replace service.sh

+++ service.sh
@@ -20,6 +20,51 @@ resetprop vendor.audio.dolby.ds2.hardbypass false
 #resetprop vendor.dolby.dap.param.tee false
 #resetprop vendor.dolby.mi.metadata.log false
 
+# function
+stop_service() {
+for NAMES in $NAME; do
+  if getprop | grep "init.svc.$NAMES\]: \[running"; then
+    stop $NAMES
+  fi
+done
+}
+run_service() {
+for FILES in $FILE; do
+  killall $FILES
+  $FILES &
+  PID=`pidof $FILES`
+done
+}
+
+# stop
+NAME="dms-hal-1-0 dms-hal-2-0 dms-v36-hal-2-0"
+stop_service
+
+# mount
+DIR=/odm/bin/hw
+FILE=$DIR/vendor.dolby_v3_6.hardware.dms360@2.0-service
+if [ "`realpath $DIR`" == $DIR ] && [ -f $FILE ]; then
+  mount -o bind $MODPATH/system/vendor/$FILE $FILE
+fi
+
+# run
+FILE="`realpath /vendor`/bin/hw/vendor.dolby.hardware.dms@2.0-service"
+run_service
+
+# restart
+VIBRATOR=`realpath /*/bin/hw/vendor.qti.hardware.vibrator.service*`
+[ "$VIBRATOR" ] && killall $VIBRATOR
+POWER=`realpath /*/bin/hw/vendor.mediatek.hardware.mtkpower@*-service`
+[ "$POWER" ] && killall $POWER
+killall android.hardware.usb@1.0-service
+killall android.hardware.usb@1.0-service.basic
+killall android.hardware.sensors@1.0-service
+killall android.hardware.sensors@2.0-service-mediatek
+killall android.hardware.light-service.mt6768
+killall android.hardware.lights-service.xiaomi_mithorium
+CAMERA=`realpath /*/bin/hw/android.hardware.camera.provider@*-service_64`
+[ "$CAMERA" ] && killall $CAMERA
+
 # wait
 sleep 20
 
@@ -30,12 +75,14 @@ if [ -d $DIR ] && [ ! -f $AML/disable ]; then
 fi
 
 # mount
-NAME="*audio*effects*.conf -o -name *audio*effects*.xml"
-#pNAME="*audio*effects*.conf -o -name *audio*effects*.xml -o -name *policy*.conf -o -name *policy*.xml"
-if [ ! -d $AML ] || [ -f $AML/disable ]; then
-  DIR=$MODPATH/system/vendor
-else
+if [ -d $AML ] && [ ! -f $AML/disable ]\
+&& find $AML/system/ -type f; then
+  NAME="*audio*effects*.conf -o -name *audio*effects*.xml"
+#p  NAME="*audio*effects*.conf -o -name *audio*effects*.xml -o -name *policy*.conf -o -name *policy*.xml"
   DIR=$AML/system/vendor
+else
+  NAME="*audio*effects*.conf -o -name *audio*effects*.xml -o -name *policy*.conf -o -name *policy*.xml"
+  DIR=$MODPATH/system/vendor
 fi
 FILE=`find $DIR/etc -maxdepth 1 -type f -name $NAME`
 if [ "`realpath /odm/etc`" == /odm/etc ] && [ "$FILE" ]; then
@@ -56,66 +103,15 @@ if [ -d /my_product/etc ] && [ "$FILE" ]; then
     fi
   done
 fi
-DIR=/odm/bin/hw
-FILE=$DIR/vendor.dolby_v3_6.hardware.dms360@2.0-service
-if [ "`realpath $DIR`" == $DIR ] && [ -f $FILE ]; then
-  mount -o bind $MODPATH/system/vendor/$FILE $FILE
+
+# restart
+PID=`pidof audioserver`
+if [ "$PID" ]; then
+  killall audioserver
 fi
-
-# restart
-killall audioserver
-
-# function
-stop_service() {
-for NAMES in $NAME; do
-  if getprop | grep "init.svc.$NAMES\]: \[running"; then
-    stop $NAMES
-  fi
-done
-}
-run_service() {
-for FILES in $FILE; do
-  killall $FILES
-  $FILES &
-  PID=`pidof $FILES`
-done
-}
-
-# stop
-NAME="dms-hal-1-0 dms-hal-2-0 dms-v36-hal-2-0
-      vendor-ozoaudio-media-c2-hal-1-0"
-stop_service
-
-# run
-FILE="`realpath /vendor`/bin/hw/vendor.dolby.hardware.dms@2.0-service
-      `realpath /vendor`/bin/hw/vendor.ozoaudio.media.c2@1.0-service"
-run_service
-
-# restart
-killall com.dolby.daxservice
-VIBRATOR=`realpath /*/bin/hw/vendor.qti.hardware.vibrator.service*`
-[ "$VIBRATOR" ] && killall $VIBRATOR
-POWER=`realpath /*/bin/hw/vendor.mediatek.hardware.mtkpower@*-service`
-[ "$POWER" ] && killall $POWER
-killall android.hardware.usb@1.0-service
-killall android.hardware.usb@1.0-service.basic
-killall android.hardware.sensors@1.0-service
-killall android.hardware.sensors@2.0-service-mediatek
-killall android.hardware.light-service.mt6768
-killall android.hardware.lights-service.xiaomi_mithorium
-CAMERA=`realpath /*/bin/hw/android.hardware.camera.provider@*-service_64`
-[ "$CAMERA" ] && killall $CAMERA
 
 # wait
 sleep 40
-
-# allow
-PKG=com.dolby.daxappui
-if pm list packages | grep $PKG ; then
-  if [ "$API" -ge 30 ]; then
-    appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
-  fi
-fi
 
 # grant
 PKG=com.dolby.daxservice
@@ -123,6 +119,15 @@ pm grant $PKG android.permission.READ_EXTERNAL_STORAGE
 pm grant $PKG android.permission.WRITE_EXTERNAL_STORAGE
 if [ "$API" -ge 30 ]; then
   appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
+fi
+killall $PKG
+
+# allow
+PKG=com.dolby.daxappui
+if pm list packages | grep $PKG ; then
+  if [ "$API" -ge 30 ]; then
+    appops set $PKG AUTO_REVOKE_PERMISSIONS_IF_UNUSED ignore
+  fi
 fi
 
 # grant
